feat: transform chat room into server component
